+++ src/Contracts/LocalizedRouter.php
@@ -2,7 +2,6 @@
 
 namespace Goodcat\L10n\Contracts;
 
-use Illuminate\Routing\Route;
 use Illuminate\Routing\RouteRegistrar;
 
 interface LocalizedRouter
@@ -13,7 +12,4 @@ interface LocalizedRouter
      * @see \Goodcat\L10n\Mixin\LocalizedRouter::lang
      */
     public function lang(array $translations = []): RouteRegistrar;
-
-    /** @see \Goodcat\L10n\Mixin\LocalizedRouter::getByKey */
-    public function getByKey(string $key): ?Route;
 }